Various changes that I don't 100% remember =/

+++ src/main/java/net/lomeli/lomlib/util/ByteUtil.java
@@ -9,15 +9,15 @@ public class ByteUtil {
 
     public static void writeIntArray(ByteBuf buffer, int... array) {
         buffer.writeInt(array.length);
-        for (int i = 0; i < array.length; i++){
+        for (int i = 0; i < array.length; i++) {
             buffer.writeInt(array[i]);
         }
     }
 
-    public static int[] readIntArray(ByteBuf buffer){
+    public static int[] readIntArray(ByteBuf buffer) {
         int[] returnArray = null;
         int size = buffer.readInt();
-        if (size > 0){
+        if (size > 0) {
             returnArray = new int[size];
             for (int i = 0; i < size; i++) {
                 returnArray[i] = buffer.readInt();
@@ -29,8 +29,8 @@ public class ByteUtil {
     public static void writeIntMap(ByteBuf buffer, HashMap<Integer, Integer> map) {
         buffer.writeInt(map.size());
         for (Map.Entry<Integer, Integer> mapEntry : map.entrySet()) {
-            int key = (Integer) mapEntry.getKey();
-            int value = (Integer) mapEntry.getValue();
+            int key = mapEntry.getKey();
+            int value = mapEntry.getValue();
             buffer.writeInt(key);
             buffer.writeInt(value);
         }
@@ -39,7 +39,7 @@ public class ByteUtil {
     public static HashMap<Integer, Integer> readIntMap(ByteBuf buffer) {
         HashMap<Integer, Integer> returnMap = new HashMap<Integer, Integer>();
         int size = buffer.readInt();
-        if (size > 0){
+        if (size > 0) {
             for (int i = 0; i < size; i++) {
                 int key = buffer.readInt();
                 int value = buffer.readInt();
